chatbot/pyproject.toml
[project]
name = "majorproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
]
